Network.AWS.S3.Encryption.Instructions

Description

Documentation

newtype Instructions

Constructors

Instructions (forall m r. (AWSConstraint r m, HasKeyEnv r) => m Envelope)

class AWSRequest a => AddInstructions a where

Minimal complete definition

add'

Methods

add' :: a -> (BucketName, ObjectKey)

Determine the bucket and key an instructions file is adjacent to.

data PutInstructions

Constructors

PutInstructions

Fields

_piExt :: Ext

_piPut :: PutObject

putInstructions :: AddInstructions a => a -> Envelope -> PutInstructions

piExtension :: Lens' PutInstructions Ext

data GetInstructions

Constructors

GetInstructions

Fields

_giExt :: Ext

_giGet :: GetObject

getInstructions :: AddInstructions a => a -> GetInstructions

giExtension :: Lens' GetInstructions Ext

class AWSRequest a => RemoveInstructions a where

Minimal complete definition

remove'

Methods

remove' :: a -> (BucketName, ObjectKey)

Determine the bucket and key an instructions file is adjacent to.

data DeleteInstructions

Constructors

DeleteInstructions

Fields

_diExt :: Ext

_diDelete :: DeleteObject

deleteInstructions :: RemoveInstructions a => a -> DeleteInstructions

diExtension :: Lens' DeleteInstructions Ext